refactor: domain, view 분
