+++ src/main/java/lotto/LottoApplication.java
@@ -1,5 +1,12 @@
 package lotto;
 
+import lotto.domain.LottoStatistics;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTicketMachine;
+import lotto.domain.PurchasePrice;
+import lotto.view.InputView;
+import lotto.view.ResultView;
+
 public class LottoApplication {
 
     public static void main(String[] args) {
