+++ src/main/java/lotto/domain/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public enum LottoRank {
     FIRST(6, false, 2000000000),
