+++ src/main/java/lotto/domain/PurchasePrice.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class PurchasePrice {
