+++ src/test/java/lotto/LottoStatisticsTest.java
@@ -3,6 +3,9 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import lotto.domain.LottoRank;
+import lotto.domain.LottoStatistics;
+import lotto.domain.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
