+++ src/test/java/lotto/LottoTicketMachineTest.java
@@ -3,6 +3,9 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTicketMachine;
+import lotto.domain.PurchasePrice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
